use waiter
this is very interesting. the waiter can be used for S3 and SQS.
basically, it can suspend the lambda function. since lambda can only run 5mins. we don't want the waiting time to count into function runtime.
for example, "PUT" may takes time to persist on S3. this will help